거래 제작 중 약간의 리팩토링

+++ Session.h
@@ -12,7 +12,7 @@
 class Session
 {
 public:
-    void Init(SOCKET socket, int index, int id, MPMCQueue<ShardMsg>* h);
+    void Init(SOCKET socket, int index, uint64_t id, MPMCQueue<ShardMsg>* h);
     void Close();
 
     void PostRecv();
